% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_check_nas.R
\name{check_column}
\alias{check_column}
\title{Checking for nas or missing data}
\usage{
check_column(working_col, test_function)
}
\arguments{
\item{working_col}{column to check for missing data}

\item{test_function}{function which returns true to identify a cell of data as "missing"}
}
\description{
The function summarizes how many values are either missing or blank and reports
that to user. Optionally, user can provide a test function which can check for additional missing values
this test_function should return true for a value if it should be flagged as missing or empty
}
