% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_check_keys.R
\name{check_keys}
\alias{check_keys}
\title{Checking join keys}
\usage{
check_keys(left, right, by = by, verbose = F, join_type = "left")
}
\arguments{
\item{left}{dataframe that we want to merge on the left side of the join}

\item{right}{dataframe that we want to merge on the right side of the join}

\item{by}{headers of columns to use to join the left and right data frames}

\item{verbose}{if true, returns more comprehensive output}

\item{join_type}{type of join ('left', right', 'inner', etc.) user is checking}
}
\description{
This function checks to see if keys being used to join are unique, as needed
for a join,and returns potential problem keys to the user.
}
