% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_content_spot_check.R
\name{content_spot_check}
\alias{content_spot_check}
\title{Spot checking overall content}
\usage{
content_spot_check(df, filename, delim = ",", rows = 15, skip = 0, verbose = T)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
