% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_encoding_check.R
\name{encoding_check}
\alias{encoding_check}
\title{Checking encoding}
\usage{
encoding_check(
  df,
 
    bad_seq_regex = "[^[`1234567890-=qwertyuiop[]\\\\asdfghjkl;'zxcvbnm,\\\\./ ~!@#$\%^&*()_+QWERTYUIOP\\\\{\\\\}|ASDFGHJKL:\\"ZXCVBNM<>?`]]"
)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
