% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_id_grouping.R
\name{id_grouping}
\alias{id_grouping}
\title{Grouping ids}
\usage{
id_grouping(df, filepath, id_column, verbose = FALSE)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
