% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_checks.R
\name{import_checks}
\alias{import_checks}
\title{Load the data checking functions as well as a high level
"run all checks" function.}
\usage{
import_checks(dataframe, filename)
}
\description{
Load the data checking functions as well as a high level
"run all checks" function.
}
