% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_leading_zero_check.R
\name{leading_zero_check}
\alias{leading_zero_check}
\title{Leading zero check}
\usage{
leading_zero_check(df, filepath)
}
\arguments{
\item{df}{data frame to check}

\item{filepath}{path to raw file we imported into data frame}
}
\description{
Function accesses columns with numeric input and checks for rows where
the numbers in columns don't match between the raw file and the dataframe,
this denotes where leading zeroes are not carried over from the raw file
to the dataframe.
}
