% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_length_match.R
\name{length_match}
\alias{length_match}
\title{Length match}
\usage{
length_match(df, filepath, skip = 0, col_names = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{data frame to check}

\item{filepath}{path to raw file we imported into data frame}

\item{skip}{how many lines at the beginning of the file to skip}

\item{col_names}{if row skipped, if this is true, the new first row is column names, if false, the new first row denotes data values}

\item{verbose}{if true, output is more comprehensive}
}
\description{
Function checks that the length of the imported data frame matches the length of
the file it was imported from.
}
