% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_multifile_load.R
\name{multifile_load}
\alias{multifile_load}
\title{Checking multifile load}
\usage{
multifile_load(
  df,
  list_of_filenames,
  skip = 0,
  col_names = TRUE,
  verbose = FALSE
)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
