% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_checks.R
\name{semi_join_dc}
\alias{semi_join_dc}
\title{Join checks}
\usage{
semi_join_dc(left, right, by = NULL, verbose = F)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
