% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_total_check.R
\name{total_check}
\alias{total_check}
\title{Total check}
\usage{
total_check(df, keywords = c("subtotal", "total", "sum"), threshold = 4)
}
\arguments{
\item{df}{dataframe we are checking}

\item{keywords}{words denoting rows with totals or sums we are trying to check for}

\item{threshold}{maximum of how many times finding the above keywords in the dataset is expected}
}
\description{
This function returns rows and columns that might be total rows, and
possibly should be separated from the dataset.
}
