% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_category_check.R
\name{category_check}
\alias{category_check}
\title{Category check}
\usage{
category_check(df, col_name)
}
\arguments{
\item{col_name}{Reference to a column to be checked for discrete categories}

\item{dataframe}{User provides reference to an R dataframe to check}
}
\description{
This function works to catch any values in the column that should be combined
with other categories. It works by catching mismatched spaces, capitalization, or punctuation
Returns list of categories that the checker suspects might overlap.
}
