% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_check_nas.R
\name{check_column}
\alias{check_column}
\title{Checking for missing data}
\usage{
check_column(working_col, test_function)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
