% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_count_rows.R
\name{count_rows}
\alias{count_rows}
\title{Checking rows in join}
\usage{
count_rows(left, right, by = NULL, join_type)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
