% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_date_check.R
\name{date_check}
\alias{date_check}
\title{Date checks}
\usage{
date_check(
  df,
  col,
  check_future = T,
  date_blacklist = as.Date(c("1899-12-31", "1900-01-01", "1901-01-01", "1904-01-02",
    "9999-09-09")),
  valid_date_range = c(1970, 2050)
)
}
\arguments{
\item{df}{pass in data frame}

\item{col}{pass in a column name}

\item{check_future}{whether to return warnings for dates in the future (default is true)}

\item{date_blacklist}{a list of dates which we don't interpret as valid}

\item{valid_date_range}{list of two years, between which dates should be interpreted as valid}
}
\description{
Function looks for dates out of the expected range, possible swapped months and dates,
or if dates are in the future
}
