% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_date_check.R
\name{date_check}
\alias{date_check}
\title{Validating dates}
\usage{
date_check(
  df,
  col,
  check_future = T,
  date_blacklist = as.Date(c("1899-12-31", "1900-01-01", "1901-01-01", "1904-01-02",
    "9999-09-09")),
  valid_date_range = c(1970, 2050)
)
}
\description{
What kind of problem are we solving?
User claims a column represents dates
We want to try to catch and see if it might have been parsed wrong
Any invalid dates
This function takes in a column that has already been assigned a Date type and
checks to see if the entries in that column make sense
}
