% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_encoding_check.R
\name{encoding_check}
\alias{encoding_check}
\title{Encoding check}
\usage{
encoding_check(
  df,
 
    bad_seq_regex = "[^[`1234567890-=qwertyuiop[]\\\\asdfghjkl;'zxcvbnm,\\\\./ ~!@#$\%^&*()_+QWERTYUIOP\\\\{\\\\}|ASDFGHJKL:\\"ZXCVBNM<>?`]]"
)
}
\arguments{
\item{df}{an R dataframe to check}

\item{bad_seq_req}{A regex of values that denote characters that should not appear
with correct encoding}
}
\description{
This function takes in a dataframe and returns rows with characters that
denote bad encoding, either using a default list, or allowing user to
provide characters to catch.
}
