% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_id_grouping_check.R
\name{id_grouping}
\alias{id_grouping}
\title{Id grouping check}
\usage{
id_grouping(df, filepath, id_column, verbose = FALSE)
}
\arguments{
\item{df}{data frame to test}

\item{filepath}{path to file that was imported, for comparison}

\item{id_column}{name of the column with ids we expect to be unique}

\item{verbose}{how extensive output of grouping function should be}
}
\description{
Function returns warning and examples of problematic ids if ids to group by are duplicated.
}
